add decorator for rpi.gpio and spidev
